model/requirements.txt
Flask==3.0.0
flask-cors==6.0.1
torch==2.8.0
torchvision==0.23.0
Pillow>=10.0.0
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
matplotlib>=3.7.2
seaborn>=0.12.2
opencv-python>=4.8.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
gevent>=23.9.0
redis>=4.5.0
Flask-Caching>=2.0.0
